+++ great_expectations/cli/upgrade_helpers/upgrade_helper_v13.py
@@ -5,7 +5,7 @@ from typing import List, Optional, Union
 
 from ruamel.yaml.comments import CommentedMap
 
-import great_expectations.data_context.toolkit as data_context_toolkit
+import great_expectations.checkpoint.toolkit as checkpoint_toolkit
 import great_expectations.exceptions as ge_exceptions
 from great_expectations import DataContext
 from great_expectations.checkpoint import Checkpoint, LegacyCheckpoint
@@ -63,7 +63,7 @@ class UpgradeHelperV13(BaseUpgradeHelper):
         self._process_validation_operators_for_checklist()
 
     def _process_checkpoint_store_for_checklist(self):
-        if data_context_toolkit.default_checkpoints_exist(
+        if checkpoint_toolkit.default_checkpoints_exist(
             directory_path=self.data_context.root_directory
         ):
             config_commented_map: CommentedMap = (
